feat: base backend (scrapes + heartbeats)
This commit includes fixes and additions that culminate in a WIP backend for the plugin:
We scrape Script Editor; in turn monitoring for the three editor events and grabbing the relevant file information.
We parse the configuration file, `~/.wakatime.cfg`, for the api_key and api_url.
We call event functions for the editor events that collectively debounce and filter heartbeat queries, ensuring appropriate data passthrough and limiting frequency.
We send heartbeats to wakatime-cli, with all of the appropriate flags.