feat: UI for permissions & configuration
This commit includes the setup flow and corresponding UI & backend for both the permissions wizard (Accessibility & Apple Events), along with the configuration (`~/.wakatime.cfg`) setup wizard (API URL & Key). This commit also includes a very rudimentary base for the main window view, relative file path obfuscation with UI, the associated sandboxing changes necessary for Apple Events permission requests, and an icon set.
The flow is the permissions wizard to the config wizard to the main view, with detection for whether or not permissions have been granted properly and whether or not the configuration file is present; skipping one or both of the wizards in the setup flow.